+++ packages/botonic-react/src/webchat/custom-messages/rating/types.ts
@@ -1,0 +1,15 @@
+export enum RatingType {
+  Stars = 'stars',
+  Smileys = 'smileys',
+}
+
+export interface RatingItemProps {
+  color: string
+  ratingNumber: number
+  hover: number
+}
+
+export interface RatingIconProps {
+  color: string
+  selected?: boolean
+}